refactor: Message 상수 위치 변경
